norm_parsing_pt2

+++ casting/load_color_from_texture.c
@@ -6,7 +6,7 @@
 /*   By: mnaqqad <mnaqqad@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/30 10:06:06 by mnaqqad           #+#    #+#             */
-/*   Updated: 2022/06/04 15:41:47 by mnaqqad          ###   ########.fr       */
+/*   Updated: 2022/06/05 11:41:55 by mnaqqad          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@ char  *load_color_from_texture(double screen_y,t_game *info)
     int texture_height = 60;
     int texture_y = screen_y * texture_height;
       if (info->side == WALL_SIDE_H)
-       texture_x = fmod(info->intersection_horizontal /(double)grid_size, 1.0) * texture_width;
+       texture_x = fmod(info->intersection_horizontal /(float)grid_size, 1.0) * texture_width;
       else
-        texture_x = fmod(info->intersection_vertical /(double)grid_size , 1.0) * texture_width;
+        texture_x = fmod(info->intersection_vertical /(float)grid_size , 1.0) * texture_width;
         return (info->addr_texture  + (texture_y * (info->line_length_texture) + texture_x * (info->bits_per_pixel_texture / 8)));
       return (NULL);
 }
